2D Platformer Godot
The 7th Expedition is a game where you play as a stranded astronaut on a faulty spaceship. To win, you must complete 7 expeditions: move forward if an anomaly is found, or move backward if none appear. Failure resets progress. Players must test objects, detect anomalies, and make the right choices to survive.